feat(service): 统一集群ID参数类型为Long，提升数据一致性
- 将多个服务类中的集群ID参数类型从Integer更改为Long，以确保与数据模型的一致性。
- 涉及的服务类包括ClusterAlertHistoryService、ClusterAlertQuotaService、ClusterServiceCommandService等，确保所有相关逻辑均能正确处理集群ID。

这些改动显著增强了服务层的健壮性与一致性，确保了整体设计的一致性。

+++ datasophon-service/src/main/java/com/datasophon/api/alert/gateway/AlertHistoryGateway.java
@@ -3,9 +3,9 @@ package com.datasophon.api.alert.gateway;
 import com.datasophon.api.alert.model.AlertHistory;
 
 public interface AlertHistoryGateway {
-    boolean hasEnabledAlertHistory(String alertname, int clusterId, String hostname);
+    boolean hasEnabledAlertHistory(String alertname, Long clusterId, String hostname);
 
-    AlertHistory getEnabledAlertHistory(String alertname, int clusterId, String hostname);
+    AlertHistory getEnabledAlertHistory(String alertname, Long clusterId, String hostname);
 
     void updateAlertHistoryToDisabled(Integer id);
